feat: refactor progress checksection import
