+++ packages/web/src/common/components/ProgressStatus/_atomic/ProgressCheckSection.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
-import { ProgressCheckSectionStatusEnum } from "../progressCheckStationStatus";
 import Progress from "./Progress";
+import { ProgressCheckSectionStatusEnum } from "./progressCheckStationStatus";
 import ProgressLine from "./ProgressLine";
 
 export interface StatusAndDate {
