+++ packages/web/src/common/components/ProgressStatus/_atomic/progressCheckStationStatus.ts
