+++ packages/web/src/features/activity-report/constants/activityReportProgress.ts
@@ -1,7 +1,7 @@
 import { ActivityStatusEnum } from "@sparcs-clubs/interface/common/enum/activity.enum";
 
-import { StatusAndDate } from "@sparcs-clubs/web/common/components/ProgressCheckSection";
-import { ProgressCheckSectionStatusEnum } from "@sparcs-clubs/web/common/components/ProgressCheckSection/progressCheckStationStatus";
+import { StatusAndDate } from "@sparcs-clubs/web/common/components/ProgressStatus/_atomic/ProgressCheckSection";
+import { ProgressCheckSectionStatusEnum } from "@sparcs-clubs/web/common/components/ProgressStatus/_atomic/progressCheckStationStatus";
 
 interface ActivityReportProgressDetail {
   labels: string[];
