+++ packages/web/src/features/manage-club/funding/constants/fundingProgressStatus.ts
@@ -1,7 +1,7 @@
 import { FundingStatusEnum } from "@sparcs-clubs/interface/common/enum/funding.enum";
 
-import { StatusAndDate } from "@sparcs-clubs/web/common/components/ProgressCheckSection";
-import { ProgressCheckSectionStatusEnum } from "@sparcs-clubs/web/common/components/ProgressCheckSection/progressCheckStationStatus";
+import { StatusAndDate } from "@sparcs-clubs/web/common/components/ProgressStatus/_atomic/ProgressCheckSection";
+import { ProgressCheckSectionStatusEnum } from "@sparcs-clubs/web/common/components/ProgressStatus/_atomic/progressCheckStationStatus";
 
 interface FundingProgressDetail {
   labels: string[];
