+++ packages/web/src/features/register-club/constants/registerClubProgress.ts
@@ -1,7 +1,7 @@
 import { RegistrationStatusEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
 
-import { StatusAndDate } from "@sparcs-clubs/web/common/components/ProgressCheckSection";
-import { ProgressCheckSectionStatusEnum } from "@sparcs-clubs/web/common/components/ProgressCheckSection/progressCheckStationStatus";
+import { StatusAndDate } from "@sparcs-clubs/web/common/components/ProgressStatus/_atomic/ProgressCheckSection";
+import { ProgressCheckSectionStatusEnum } from "@sparcs-clubs/web/common/components/ProgressStatus/_atomic/progressCheckStationStatus";
 
 interface RegisterClubProgressDetail {
   labels: string[];
